chore: removes comments and separates constants into a seperate file
